enhance: code style of mdx-compiler
