+++ fixtures/workspace/packages/library/components/Button/__docs__/index.mdx
@@ -3,6 +3,8 @@ description: Basic button ui component.
 labels: ['button', 'ui', 'common']
 ---
 
+![](./test.png)
+
 import Basic from './basic';
 import Advanced from './advanced';
