+++ packages/aspect/src/mdx/loader/remarkPlugins/index.ts
@@ -1,0 +1,4 @@
+export { extractComponentDemos } from './extractComponentDemos';
+export { extractHeadings } from './extractHeadings';
+export { extractImports } from './extractImports';
+export { extractMetadata } from './extractMetadata';
